Factor out interaction code into a module
